Factor out docker_django_management.py.
