+++ docker-compose.yml
@@ -5,9 +5,10 @@ app:
   links:
     - db
   working_dir: /calc
-  entrypoint: python /calc/entrypoint.py
+  entrypoint: python /calc/docker_django_management.py
   environment:
-    - IS_RUNNING_IN_DOCKER=yup
+    - DDM_HOST_USER=calc_user
+    - DDM_IS_RUNNING_IN_DOCKER=yup
     - PYTHONUNBUFFERED=yup
     - DATABASE_URL=postgres://calc_user@db/calc
   command: python manage.py runserver 0.0.0.0:8000
